move local storage management to isolate file
