+++ src/main.tsx
@@ -2,9 +2,10 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
+import {LocalStorage} from "./services/localStorage.tsx";
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
-    <App />
+    <App storage={LocalStorage()} />
   </React.StrictMode>,
 )
